Criação da página de códigos; adição de animações com JQuery e ajuste dos formulários de login e registro.

+++ database/migrations/2024_09_19_182933_alter_report_remove_colum_lastename.php
@@ -14,7 +14,9 @@ class AlterReportRemoveColumLastename extends Migration
     public function up()
     {
         Schema::table('reports', function (Blueprint $table) {
-            $table->dropColumn('lastname');
+            if (Schema::hasColumn('reports', 'lastname')) {
+                $table->dropColumn('lastname');
+            }
         });
     }
